Simple Dialogue System + small Mana bar fix
Added a "Dialogue" folder with a test scene and some scripts. Press "Start Dialogue" for text to appear. See "DialogueBox" prefab to change what text the dialogue should be (expand "Dialogue").
Also changed places for the "mpText" line in the manabar script because otherwise, that line does nothing ("return;" is run right before)